Refactor(NumberRounder): 소숫점 첫째 자리수 변환 객체 생성
