+++ src/main/java/lotto/model/service/prizecalculator/NumberRounder.java
@@ -1,0 +1,8 @@
+package lotto.model.service.prizecalculator;
+
+public class NumberRounder {
+
+    public double roundTenthsPlaceValue(double profitRate) {
+        return Math.round(profitRate * 10) / 10.0;
+    }
+}
